+++ src/TestUtils/TestDatasystemProviderService.php
@@ -11,64 +11,84 @@ use Symfony\Component\HttpFoundation\Response;
 
 class TestDatasystemProviderService implements DatasystemProviderServiceInterface
 {
-    private array $data;
+    private static ?array $data = null;
 
     public function __construct()
     {
-        $this->data = [];
+        if (self::$data !== null) {
+            self::cleanup();
+        }
+        self::$data = [];
     }
 
     public function hasFile(string $internalBucketId, string $fileId): bool
     {
-        return isset($this->data[$internalBucketId][$fileId]);
+        return isset(self::$data[$internalBucketId][$fileId]);
     }
 
     public function saveFile(string $internalBucketId, string $fileId, File $file): void
     {
-        $this->data[$internalBucketId][$fileId] = $file;
+        if (isset(self::$data[$internalBucketId][$fileId])) {
+            @unlink(self::$data[$internalBucketId][$fileId]->getRealPath());
+        }
+
+        $tempFilePath = tempnam(sys_get_temp_dir(), 'dbp_relay_blob_bundle_unit_test_tempfile_');
+        copy($file->getRealPath(), $tempFilePath);
+
+        self::$data[$internalBucketId][$fileId] = new File($tempFilePath);
     }
 
     public function getBinaryResponse(string $internalBucketId, string $fileId): Response
     {
-        if (!isset($this->data[$internalBucketId][$fileId])) {
+        if (!isset(self::$data[$internalBucketId][$fileId])) {
             throw new \RuntimeException();
         }
 
-        // build binary response
-        $response = new BinaryFileResponse($this->data[$internalBucketId][$fileId]->getRealPath());
-
-        return $response;
+        return new BinaryFileResponse(self::$data[$internalBucketId][$fileId]->getRealPath());
     }
 
     public function removeFile(string $internalBucketId, string $fileId): void
     {
-        if (!isset($this->data[$internalBucketId][$fileId])) {
+        if (!isset(self::$data[$internalBucketId][$fileId])) {
             throw new \RuntimeException();
         }
 
-        unset($this->data[$internalBucketId][$fileId]);
+        @unlink(self::$data[$internalBucketId][$fileId]->getRealPath());
+        unset(self::$data[$internalBucketId][$fileId]);
     }
 
     public function listFiles(string $internalBucketId): iterable
     {
-        return array_keys($this->data[$internalBucketId] ?? []);
+        return array_keys(self::$data[$internalBucketId] ?? []);
     }
 
     public function getFileSize(string $internalBucketId, string $fileId): int
     {
-        if (!isset($this->data[$internalBucketId][$fileId])) {
+        if (!isset(self::$data[$internalBucketId][$fileId])) {
             throw new \RuntimeException();
         }
 
-        return $this->data[$internalBucketId][$fileId]->fileSize();
+        return self::$data[$internalBucketId][$fileId]->fileSize();
     }
 
     public function getFileHash(string $internalBucketId, string $fileId): string
     {
-        if (!isset($this->data[$internalBucketId][$fileId])) {
+        if (!isset(self::$data[$internalBucketId][$fileId])) {
             throw new \RuntimeException();
         }
 
-        return hash('sha256', $this->data[$internalBucketId][$fileId]->getContent());
+        return hash('sha256', self::$data[$internalBucketId][$fileId]->getContent());
+    }
+
+    public static function cleanup(): void
+    {
+        if (self::$data !== null) {
+            foreach (self::$data as $files) {
+                foreach ($files as $file) {
+                    @unlink($file->getRealPath());
+                }
+            }
+        }
+        self::$data = null;
     }
 }